fix:메인페이지에서 설정 아이콘 삭제, 친구추가 아이콘 친구 페이지로 이동,메뉴바에서 기록과 마이맵 탭 삭제 #111
